PLCON-68 VPN Coexistence Linux - adding logic both for nftables and iptables-legacy. Initial pass, stubbed out iptables-legacy path.
